requirements.txt
streamlit
joblib
xgboost
numpy
scikit-learn
pillow